🐛 avoid to execute view tests twice
